feat: implement complete continuous training integration
- Add ContinuousTrainingClient for AI service communication
- Add ScheduledTrainingService for automatic training triggers
- Update DiagnosisController with training endpoints and feedback integration
- Fix DTO field types (Double -> Float) for AI service compatibility
- Add frontend accept/reject buttons with feedback API calls
- Add comprehensive training integration test script
- Enable scheduling in GenAIServiceApplication

Features:
- Real-time feedback counting and threshold detection
- Automatic training triggers at 100 feedback entries
- Manual training triggers and eligibility checks
- Frontend integration for doctor feedback collection
- Training job status monitoring and completion tracking
- Model version updates and deployment tracking

+++ spring-petclinic-genai-service/src/main/java/org/springframework/samples/petclinic/genai/dto/AiDiagnosisRequest.java
@@ -4,8 +4,8 @@ public record AiDiagnosisRequest(
         String animal_type,
         String gender,
         Integer age_months,
-        Double weight_kg,
-        Double temperature,
+        Float weight_kg,           // ✅ Changed to Float
+        Float temperature,         // ✅ Changed to Float
         Integer heart_rate,
         String current_season,
         String vaccination_status,
